backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
redis==5.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
scikit-learn==1.3.2
xgboost==2.0.3
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2
python-multipart==0.0.6
python-dotenv==1.0.0

